+++ CPP Module 05/ex00/Bureaucrat.hpp
@@ -7,17 +7,17 @@
 class Bureaucrat {
     private:
         const std::string _name;
-        size_t _grade;
+        unsigned int _grade;
         void setGrade(int n);
     public:
         Bureaucrat(void);
         ~Bureaucrat(void);
         Bureaucrat(std::string name);
-        Bureaucrat(std::string name, size_t grade);
+        Bureaucrat(std::string name, int grade);
         Bureaucrat(const Bureaucrat& copy);
         Bureaucrat &operator=(const Bureaucrat& copy);
         std::string getName(void) const;
-        size_t getGrade(void) const;
+        unsigned int getGrade(void) const;
         void increaseGrade(void);
         void decreaseGrade(void);
         class GradeTooHighException: public std::exception {
